Finally fixed Draw_Sprite fully
When I instructed Draw_Sprite to return whenever it first encounters a null pointer, I did not consider that a byte of hex value 0x0 is considered null. Instead it terminates when it reaches the end of xMax * yMax, which finally made the not rendering bug go away completely from what I could see.